langgraph/06-observability/requirements.txt
# LangSmith observability
langsmith>=0.1.0

# LangGraph (for running the examples we instrument)
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0

# Environment variables
python-dotenv==1.0.0
